+++ ASPNETCore2JwtAuthentication.sln
@@ -17,6 +17,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ASPNETCore2JwtAuthenticatio
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ASPNETCore2JwtAuthentication.ConsoleClient", "src\ASPNETCore2JwtAuthentication.ConsoleClient\ASPNETCore2JwtAuthentication.ConsoleClient.csproj", "{47536CB0-1B7E-4089-AFBF-D7C55A6076EC}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ASPNETCore2JwtAuthentication.IoCConfig", "src\ASPNETCore2JwtAuthentication.IoCConfig\ASPNETCore2JwtAuthentication.IoCConfig.csproj", "{9C3184B8-F3AC-4ED9-A352-C861AF264461}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -102,6 +104,18 @@ Global
 		{47536CB0-1B7E-4089-AFBF-D7C55A6076EC}.Release|x64.Build.0 = Release|Any CPU
 		{47536CB0-1B7E-4089-AFBF-D7C55A6076EC}.Release|x86.ActiveCfg = Release|Any CPU
 		{47536CB0-1B7E-4089-AFBF-D7C55A6076EC}.Release|x86.Build.0 = Release|Any CPU
+		{9C3184B8-F3AC-4ED9-A352-C861AF264461}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9C3184B8-F3AC-4ED9-A352-C861AF264461}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9C3184B8-F3AC-4ED9-A352-C861AF264461}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{9C3184B8-F3AC-4ED9-A352-C861AF264461}.Debug|x64.Build.0 = Debug|Any CPU
+		{9C3184B8-F3AC-4ED9-A352-C861AF264461}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{9C3184B8-F3AC-4ED9-A352-C861AF264461}.Debug|x86.Build.0 = Debug|Any CPU
+		{9C3184B8-F3AC-4ED9-A352-C861AF264461}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9C3184B8-F3AC-4ED9-A352-C861AF264461}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9C3184B8-F3AC-4ED9-A352-C861AF264461}.Release|x64.ActiveCfg = Release|Any CPU
+		{9C3184B8-F3AC-4ED9-A352-C861AF264461}.Release|x64.Build.0 = Release|Any CPU
+		{9C3184B8-F3AC-4ED9-A352-C861AF264461}.Release|x86.ActiveCfg = Release|Any CPU
+		{9C3184B8-F3AC-4ED9-A352-C861AF264461}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{34852B0F-2494-43F0-B120-8918D0A97DD7} = {EB31F3BE-2CB7-4C42-8E35-FF97075ECB4E}
@@ -110,5 +124,6 @@ Global
 		{75DD2312-9B4F-497A-9FBE-8377C51B1E79} = {EB31F3BE-2CB7-4C42-8E35-FF97075ECB4E}
 		{34E2F813-5901-4069-9515-BFEB39DE3222} = {EB31F3BE-2CB7-4C42-8E35-FF97075ECB4E}
 		{47536CB0-1B7E-4089-AFBF-D7C55A6076EC} = {EB31F3BE-2CB7-4C42-8E35-FF97075ECB4E}
+		{9C3184B8-F3AC-4ED9-A352-C861AF264461} = {EB31F3BE-2CB7-4C42-8E35-FF97075ECB4E}
 	EndGlobalSection
 EndGlobal
