+++ global.json
@@ -1,5 +1,5 @@
 {
   "sdk": {
-    "version": "3.1.102"
+    "version": "3.1.201"
   }
 }
