+++ src/ASPNETCore2JwtAuthentication.DataLayer/_01-add_migrations.cmd
@@ -1,5 +1,5 @@
-dotnet tool install --global dotnet-ef --version 3.1.2
-dotnet tool update --global dotnet-ef --version 3.1.2
+dotnet tool install --global dotnet-ef --version 3.1.3
+dotnet tool update --global dotnet-ef --version 3.1.3
 dotnet build
 For /f "tokens=2-4 delims=/ " %%a in ('date /t') do (set mydate=%%c_%%a_%%b)
 For /f "tokens=1-2 delims=/:" %%a in ("%TIME: =0%") do (set mytime=%%a%%b)
