+++ src/ASPNETCore2JwtAuthentication.DataLayer/_02-update_db.cmd
@@ -1,5 +1,5 @@
-dotnet tool install --global dotnet-ef --version 3.1.2
-dotnet tool update --global dotnet-ef --version 3.1.2
+dotnet tool install --global dotnet-ef --version 3.1.3
+dotnet tool update --global dotnet-ef --version 3.1.3
 dotnet build
 dotnet ef --startup-project ../ASPNETCore2JwtAuthentication.WebApp/ database update
 pause
