+++ src/ASPNETCore2JwtAuthentication.WebApp/Startup.cs
@@ -1,24 +1,15 @@
-﻿using System;
-using System.IO;
-using System.Text;
+﻿using System.IO;
 using System.Text.Json;
-using System.Threading.Tasks;
-using ASPNETCore2JwtAuthentication.DataLayer.Context;
 using ASPNETCore2JwtAuthentication.Services;
-using ASPNETCore2JwtAuthentication.WebApp.Models;
-using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Diagnostics;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
 using Microsoft.IdentityModel.Tokens;
 using Microsoft.Extensions.Hosting;
-
+using ASPNETCore2JwtAuthentication.IoCConfig;
 
 namespace ASPNETCore2JwtAuthentication.WebApp
 {
@@ -34,113 +25,12 @@ namespace ASPNETCore2JwtAuthentication.WebApp
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddOptions<BearerTokensOptions>()
-                    .Bind(Configuration.GetSection("BearerTokens"))
-                    .Validate(bearerTokens =>
-                    {
-                        return bearerTokens.AccessTokenExpirationMinutes < bearerTokens.RefreshTokenExpirationMinutes;
-                    }, "RefreshTokenExpirationMinutes is less than AccessTokenExpirationMinutes. Obtaining new tokens using the refresh token should happen only if the access token has expired.");
-            services.AddOptions<ApiSettings>()
-                    .Bind(Configuration.GetSection("ApiSettings"));
-
-            services.AddSingleton<IHttpContextAccessor, HttpContextAccessor>();
-            services.AddScoped<IAntiForgeryCookieService, AntiForgeryCookieService>();
-            services.AddScoped<IUnitOfWork, ApplicationDbContext>();
-            services.AddScoped<IUsersService, UsersService>();
-            services.AddScoped<IRolesService, RolesService>();
-            services.AddSingleton<ISecurityService, SecurityService>();
-            services.AddScoped<IDbInitializerService, DbInitializerService>();
-            services.AddScoped<ITokenStoreService, TokenStoreService>();
-            services.AddScoped<ITokenValidatorService, TokenValidatorService>();
-            services.AddScoped<ITokenFactoryService, TokenFactoryService>();
-
-            services.AddDbContext<ApplicationDbContext>(options =>
-            {
-                options.UseSqlServer(
-                    Configuration.GetConnectionString("DefaultConnection")
-                                .Replace("|DataDirectory|", Path.Combine(Directory.GetCurrentDirectory(), "wwwroot", "app_data")),
-                    serverDbContextOptionsBuilder =>
-                        {
-                            var minutes = (int)TimeSpan.FromMinutes(3).TotalSeconds;
-                            serverDbContextOptionsBuilder.CommandTimeout(minutes);
-                            serverDbContextOptionsBuilder.EnableRetryOnFailure();
-                        });
-            });
-
-            // Only needed for custom roles.
-            services.AddAuthorization(options =>
-                    {
-                        options.AddPolicy(CustomRoles.Admin, policy => policy.RequireRole(CustomRoles.Admin));
-                        options.AddPolicy(CustomRoles.User, policy => policy.RequireRole(CustomRoles.User));
-                        options.AddPolicy(CustomRoles.Editor, policy => policy.RequireRole(CustomRoles.Editor));
-                    });
-
-            // Needed for jwt auth.
-            services
-                .AddAuthentication(options =>
-                {
-                    options.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
-                    options.DefaultSignInScheme = JwtBearerDefaults.AuthenticationScheme;
-                    options.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
-                })
-                .AddJwtBearer(cfg =>
-                {
-                    cfg.RequireHttpsMetadata = false;
-                    cfg.SaveToken = true;
-                    cfg.TokenValidationParameters = new TokenValidationParameters
-                    {
-                        ValidIssuer = Configuration["BearerTokens:Issuer"], // site that makes the token
-                        ValidateIssuer = false, // TODO: change this to avoid forwarding attacks
-                        ValidAudience = Configuration["BearerTokens:Audience"], // site that consumes the token
-                        ValidateAudience = false, // TODO: change this to avoid forwarding attacks
-                        IssuerSigningKey = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(Configuration["BearerTokens:Key"])),
-                        ValidateIssuerSigningKey = true, // verify signature to avoid tampering
-                        ValidateLifetime = true, // validate the expiration
-                        ClockSkew = TimeSpan.Zero // tolerance for the expiration date
-                    };
-                    cfg.Events = new JwtBearerEvents
-                    {
-                        OnAuthenticationFailed = context =>
-                        {
-                            var logger = context.HttpContext.RequestServices.GetRequiredService<ILoggerFactory>().CreateLogger(nameof(JwtBearerEvents));
-                            logger.LogError("Authentication failed.", context.Exception);
-                            return Task.CompletedTask;
-                        },
-                        OnTokenValidated = context =>
-                        {
-                            var tokenValidatorService = context.HttpContext.RequestServices.GetRequiredService<ITokenValidatorService>();
-                            return tokenValidatorService.ValidateAsync(context);
-                        },
-                        OnMessageReceived = context =>
-                        {
-                            return Task.CompletedTask;
-                        },
-                        OnChallenge = context =>
-                        {
-                            var logger = context.HttpContext.RequestServices.GetRequiredService<ILoggerFactory>().CreateLogger(nameof(JwtBearerEvents));
-                            logger.LogError("OnChallenge error", context.Error, context.ErrorDescription);
-                            return Task.CompletedTask;
-                        }
-                    };
-                });
-
-
-            services.AddCors(options =>
-            {
-                options.AddPolicy("CorsPolicy",
-                    builder => builder
-                        .WithOrigins("http://localhost:4200") //Note:  The URL must be specified without a trailing slash (/).
-                        .AllowAnyMethod()
-                        .AllowAnyHeader()
-                        .SetIsOriginAllowed((host) => true)
-                        .AllowCredentials());
-            });
-
-            services.AddAntiforgery(x => x.HeaderName = "X-XSRF-TOKEN");
-            services.AddMvc(options =>
-            {
-                options.Filters.Add(new AutoValidateAntiforgeryTokenAttribute());
-            });
+            services.AddCustomOptions(Configuration);
+            services.AddCustomServices();
+            services.AddCustomDbContext(Configuration);
+            services.AddCustomJwtBearer(Configuration);
+            services.AddCustomCors();
+            services.AddCustomAntiforgery();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
@@ -165,7 +55,7 @@ namespace ASPNETCore2JwtAuthentication.WebApp
                 appBuilder.Use(async (context, next) =>
                 {
                     var error = context.Features[typeof(IExceptionHandlerFeature)] as IExceptionHandlerFeature;
-                    if (error != null && error.Error is SecurityTokenExpiredException)
+                    if (error?.Error is SecurityTokenExpiredException)
                     {
                         context.Response.StatusCode = 401;
                         context.Response.ContentType = "application/json";
@@ -175,7 +65,7 @@ namespace ASPNETCore2JwtAuthentication.WebApp
                             Msg = "token expired"
                         }));
                     }
-                    else if (error != null && error.Error != null)
+                    else if (error?.Error != null)
                     {
                         context.Response.StatusCode = 500;
                         context.Response.ContentType = "application/json";
