+++ tag-it.bat
@@ -1,3 +1,3 @@
-git tag -a 3.1.102 -m "Published V3.1.102"
+git tag -a 3.1.201 -m "Published V3.1.201"
 git push --follow-tags
 pause
